api/requirements.txt
tornado>=6.4
google-genai>=1.0
pydantic>=2.0

